Leaver user’s PATH untouched.
Instead of adding the location of qfc to the user’s PATH, this
leaves PATH untouched and calls qfc at an arbitrary install
location instead.

In addition to not adding noise to the user’s path for just one
script, this also allows having qfc installed to a location other
than ~/.qfc.